pyproject.toml
[tool.poetry]
name = "github_analytics_platform"
version = "0.1.0"
description = "A real-time analytics platform for the GitHub universe using PySpark."
authors = ["shardulkumar@live.in"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyspark = "^3.5.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-spark = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
